backend/requirements.txt
Flask==3.0.2
flask-mongoengine-3==1.1.0
pymongo[srv]==4.6.2
python-dotenv==1.0.1
pytest==8.0.2
flask-cors==4.0.0
dnspython==2.6.1
